Adds the friendship bracelet. Adds additional options for debugging (press d on any applet).

+++ animations/src/com/resamsel/bluedots/BlueDots.java
@@ -100,8 +100,8 @@ public class BlueDots extends InteractiveExportPApplet {
 	}
 
 	@Override
-	protected float getMaxFrameCount() {
-		return halfDiagonal / waveAdvance;
+	protected int getMaxFrameCount() {
+		return (int) (halfDiagonal / waveAdvance);
 	}
 
 	@Override
