+++ animations/src/com/resamsel/starpattern/StarPattern.java
@@ -154,8 +154,8 @@ public class StarPattern extends InteractiveExportPApplet {
 	 * @see com.resamsel.ExportPApplet#getMaxFrameCount()
 	 */
 	@Override
-	protected float getMaxFrameCount() {
-		return maxAngle / angleAdvance;
+	protected int getMaxFrameCount() {
+		return (int) (maxAngle / angleAdvance);
 	}
 
 	/**
